fixed: 쓰기 문제를 발견하고 수정완료
decoder에서 읽을 때 바이트를 잘 못 파싱하던 것이 문제였다.

+++ ModbusClient/src/main/java/com/example/ModbusClient/dto/modbus/WriteRequestParameters.java
@@ -10,6 +10,6 @@ import java.util.List;
 public class WriteRequestParameters {
     private int parameterCount;
     private int startAddress;
-    private double value;
-    private List<Double> values;
+    private int value;
+    private List<Integer> values;
 }
